creation and deletion of the user done

+++ src/main/java/com/example/demo/model/orm/User.java
@@ -33,7 +33,7 @@ public class User {
 
     @Column(name = "ROLE_ID",
 	    nullable = false)
-    private Long roleId;
+    private int roleId;
 
     @Column(name = "Password",
 	    nullable = false)
@@ -63,7 +63,8 @@ public class User {
     @Column(name = "Birthday")
     @Temporal(TemporalType.DATE)
     private Date birthday;
+    
     @Transient
-    private long loginId;
+    private int loginId;
 
 }
